+++ scripts/render_host_api_docs.sh
@@ -13,4 +13,5 @@ GIT_ROOT=$(git rev-parse --show-toplevel)
     make clean html
 )
 
+echo http://localhost:9000
 python3 -m http.server 9000 -d "${GIT_ROOT}/docs/build"