+++ changelog.md
@@ -5,6 +5,9 @@
 
 ## New features
 
+* Added a new utility function to color relationships `VisualizationGraph.color_relationships`
+* Added a new utility function to resize relationships `VisualizationGraph.resize_relationships`
+
 
 ## Bug fixes